add firebase integration, error handling

+++ src/main/java/com/postgresql/Daruma/repo/SkinRepo.java
@@ -1,10 +1,10 @@
 package com.postgresql.Daruma.repo;
 
-import com.postgresql.Daruma.model.Skin;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.postgresql.Daruma.model.Skin;
+
 @Repository
-public interface SkinRepo extends JpaRepository<Skin, Long> {
+public interface SkinRepo extends JpaRepository<Skin, Integer> {
 }
